Moved connect logic to another file alongside other queries and future queries, tested
