+++ services/passport.js
@@ -3,7 +3,7 @@ var GoogleStrategy = require('passport-google-oauth20').Strategy;
 var keys = require('../config/keys');
 var mongoose = require('mongoose');
 var User = mongoose.model('users');
-var db = require('../config/db-connection');
+var db = require('../database/queries');
 
 passport.use(
     new GoogleStrategy({
@@ -15,6 +15,9 @@ passport.use(
         // Once the user has OAuth'ed with Google, we retrieve the access token and save them as a User model in our DB
         var existingUser = await User.findOne({ googleId: profile.id });
 
+        db.insertUser();
+
+        
         if (existingUser) {
             return done(null, existingUser); // Tells passport we're done w/ this user model
         }
